Seperated Line chart from list session
