+++ Maestro/src/pages/home/home.html
@@ -18,5 +18,4 @@
     <h3 text-center>Welcome</h3>
 
     <list-session [amount]="5" (isLoaded)="areSessionsLoaded($event)" [showGraph]="true"></list-session>
-
 </ion-content>
