+++ Maestro/src/shared/survey_temp/survey_temp.scss
